mission2/code/ramps_X_stepper/pyproject.toml
[project]
name = "ramps-x-stepper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pyserial>=3.5", "pynput>=1.7.6"]
